Refactor: Remove sidebar, add NewAudioDialog component
